extract log lib
